change int types of parameters for EmitDBStateEventsFromHeight and GetAllDBStateMsgsFromDatabase so no conversion is needed

+++ common/interfaces/state.go
@@ -334,10 +334,10 @@ type IState interface {
 	GetIgnoreDone() bool
 
 	// Emit DBState events to the livefeed api from a specified height
-	EmitDBStateEventsFromHeight(height int64, end int64)
+	EmitDBStateEventsFromHeight(height uint32, end uint32)
 
 	// Return all DBStateMsgs from a specified height
-	GetAllDBStateMsgsFromDatabase(height int64, end int64) []IMsg
+	GetAllDBStateMsgsFromDatabase(height uint32, end uint32) []IMsg
 
 	// Access to Holding Queue
 	LoadHoldingMap() map[[32]byte]IMsg
